refactor(settings): moved steam credentials section to its own settings tab
