+++ docs/content/docs/settings/general.mdx
@@ -51,10 +51,6 @@ Choose the currency you want to use when displaying prices in SGI, such as for [
 
 This currency will also be used to determine the sell price for trading cards
 
-## Steam Credentials
-
-SGI needs your Steam community credentials for the account you want to farm trading cards on. This allows it to determine which games still have card drops available. Check the [Steam Credentials](/docs/steam-credentials) section for more information
-
 ## Steam Web API Key
 Allows users to provide their own Steam Web API key to fetch profile and game data. API keys are free for all users with a Steam account, you can [get yours here](https://steamcommunity.com/dev/apikey)
   - Useful if users want to keep their Steam profile and game details set to private
