+++ docs/content/docs/steam-credentials.mdx
@@ -58,7 +58,7 @@ This allows SGI to verify which of your games have card drops remaining, as well
 
     <Steps>
       <Step>
-        Open SGI and go to [<MockButton type='context-cog' content='Settings → General → Steam Credentials' />](/docs/settings/general#steam-credentials)
+        Open SGI and go to [<MockButton type='context-cog' content='Settings → Steam Credentials' />](/docs/steam-credentials)
       </Step>
 
       <Step>
