+++ src/components/settings/Settings.tsx
@@ -11,6 +11,7 @@ import CardSettings from '@/components/settings/CardSettings'
 import GameSettings from '@/components/settings/GameSettings'
 import GeneralSettings from '@/components/settings/GeneralSettings'
 import Logs from '@/components/settings/Logs'
+import SteamCredentials from '@/components/settings/SteamCredentials'
 import TradingCardManagerSettings from '@/components/settings/TradingCardManagerSettings'
 import ViewDocumentation from '@/components/settings/ViewDocumentation'
 import AdSlot from '@/components/ui/AdSlot'
@@ -71,6 +72,9 @@ export default function Settings(): ReactElement {
         </Tab>
         <Tab key='trading-card-manager' title={t('tradingCards.title')}>
           <TradingCardManagerSettings />
+        </Tab>
+        <Tab key='steam-credentials' title={t('settings.cardFarming.steamCredentialsTitle')}>
+          <SteamCredentials />
         </Tab>
         <Tab key='game-settings' title={t('common.gameSettings')}>
           <GameSettings />
