+++ src/i18n/locales/en-US/translation.json
@@ -159,13 +159,13 @@
   "settings.general.webApi.description": "Use your own Steam web API key. <1>Learn more</1>",
   "settings.general.webApi.descriptionTwo": "Get your Steam Web API key from <1>https://steamcommunity.com/dev/apikey</1>",
   "settings.general.helpTranslate": "Help translate SGI. <1>Learn more</1>",
-  "settings.cardFarming.alert": "To use the Card Farming feature, please add your Steam Credentials in Settings → General → Steam Credentials",
+  "settings.cardFarming.alert": "To use the Card Farming feature, please add your Steam Credentials in Settings → Steam Credentials",
   "settings.cardFarming.listGames": "Farm Games In List",
   "settings.cardFarming.listGames.description": "Only farm games that are in your Card Farming list",
   "settings.cardFarming.allGames": "Farm All Games",
   "settings.cardFarming.allGames.description": "Farm cards for all games in your library that have card drops remaining",
   "settings.cardFarming.steamCredentialsTitle": "Steam Credentials",
-  "settings.cardFarming.steamCredentials": "Steam credentials are required in order to use the Card Farming feature. <1>Learn more</1>",
+  "settings.cardFarming.steamCredentials": "Steam credentials are required in order to use the Card Farming and Trading Card Manager features. <1>Learn more</1>",
   "settings.cardFarming.steamCredentialsTwo": "Get your Steam credentials from <1>https://steamcommunity.com</1>",
   "settings.cardFarming.gamesWithDrops": "Games with drops:",
   "settings.cardFarming.totalDrops": "Total drops remaining:",
@@ -181,7 +181,7 @@
   "settings.achievementUnlocker.unlockInterval": "Unlock Interval",
   "settings.achievementUnlocker.interval": "Unlock achievements randomly every {{min}} to {{max}} minutes",
   "settings.achievementUnlocker.nextTask.description": "Run the next task automatically after Achievement Unlocker is complete",
-  "settings.tradingCards.alert": "To use the Trading Card Manager feature, please add your Steam Credentials in Settings → General → Steam Credentials",
+  "settings.tradingCards.alert": "To use the Trading Card Manager feature, please add your Steam Credentials in Settings → Steam Credentials",
   "settings.tradingCards.sellOptions": "Default Sell Option",
   "settings.tradingCards.sellOptions.description": "Choose the default sell price for trading cards when using the 'Sell All' button",
   "settings.tradingCards.sellOptions.highestBuyOrder": "Highest Buy Order",
@@ -221,8 +221,8 @@
   "proMode.removeAdsWith": "Remove ads with",
   "toast.steam": "Steam is not running",
   "toast.mismatch": "Account mismatch between Steam and SGI",
-  "toast.missingCredentials": "Missing card farming credentials in \"Settings → General → Steam Credentials\"",
-  "toast.outdatedCredentials": "Card farming credentials need to be updated in \"Settings → General → Steam Credentials\"",
+  "toast.missingCredentials": "Missing card farming credentials in \"Settings → Steam Credentials\"",
+  "toast.outdatedCredentials": "Card farming credentials need to be updated in \"Settings → Steam Credentials\"",
   "toast.incorrectCredentials": "Incorrect card farming credentials",
   "toast.enableAllGames": "Add some games to your card farming list or enable \"all games\" in \"Settings → Card Farming\"",
   "toast.noGames": "There are no games in your achievement unlocker list",
